// src/types/enums.ts
export enum SceneKey {
  Boot = 'Boot',
  Preload = 'Preload',
  MainMenu = 'MainMenu',
  Town = 'Town',
  Shop = 'Shop',
  Hero = 'Hero',
  Dungeon = 'Dungeon',
  MicroGameSelect = 'MicroGameSelect',
  Match3 = 'Match3',
  MemoryMatch = 'MemoryMatch',
  SimonSays = 'SimonSays',
  Prestige = 'Prestige',
  Settings = 'Settings',
  HUD = 'HUD',
}

export enum HeroClass {
  Warrior = 'Warrior',
  Mage = 'Mage',
  Ranger = 'Ranger',
  Cleric = 'Cleric',
  Thief = 'Thief',
}

export enum ItemSlotType {
  Weapon = 'Weapon',
  Armor = 'Armor',
  Accessory = 'Accessory',
}

export enum ItemRarity {
  Common = 'Common',
  Uncommon = 'Uncommon',
  Rare = 'Rare',
  Epic = 'Epic',
  Legendary = 'Legendary',
}

export enum UpgradeCategory {
  Click = 'Click',
  Auto = 'Auto',
  Hero = 'Hero',
  Equipment = 'Equipment',
  Prestige = 'Prestige',
}

export enum MicroGameType {
  Match3 = 'Match3',
  MemoryMatch = 'MemoryMatch',
  SimonSays = 'SimonSays',
}

export enum CombatAction {
  Attack = 'Attack',
  Skill = 'Skill',
  Defend = 'Defend',
}

export enum EnemyType {
  Normal = 'Normal',
  Elite = 'Elite',
  Boss = 'Boss',
}

export enum ShopTab {
  Upgrades = 'Upgrades',
  Equipment = 'Equipment',
  Potions = 'Potions',
  Party = 'Party',
}